forgot to commit?

+++ nfs_ops.go
@@ -102,7 +102,7 @@ func (nfs *Nfs) Lookup(args *LOOKUP3args, reply *LOOKUP3res) error {
 func (nfs *Nfs) Access(args *ACCESS3args, reply *ACCESS3res) error {
 	log.Printf("Access %v\n", args)
 	reply.Status = NFS3_OK
-	reply.Resok.Access = ACCESS3_READ | ACCESS3_LOOKUP | ACCESS3_MODIFY | ACCESS3_EXTEND | ACCESS3_DELETE | ACCESS3_EXECUTE
+	reply.Resok.Access = Uint32(ACCESS3_READ | ACCESS3_LOOKUP | ACCESS3_MODIFY | ACCESS3_EXTEND | ACCESS3_DELETE | ACCESS3_EXECUTE)
 	return nil
 }
